Sensors/WaterLinkedDVL: extract Driver
