+++ src/Sensors/WaterLinkedDVL/Task.cpp
@@ -34,8 +34,9 @@
 // DUNE headers.
 #include <DUNE/DUNE.hpp>
 
-// JSON headers.
-#include <json.hpp>
+// Local Headers
+#include "Configuration.hpp"
+#include "Driver.hpp"
 
 namespace Sensors
 {
@@ -44,14 +45,9 @@ namespace Sensors
   namespace WaterLinkedDVL
   {
     using DUNE_NAMESPACES;
-    using json = nlohmann::json;
 
-    //! Number of DVL beams.
-    static const unsigned c_beams = 4;
     //! Transducer beam offset (the distance between the center of transducers and the device center).
     static const float c_xdcr_offset = 0.02;
-    //! Command ticket time to live in seconds.
-    static const float c_ttl = 1.0;
 
     //! %Task arguments.
     struct Arguments
@@ -74,61 +70,35 @@ namespace Sensors
 
     struct Task : public Hardware::BasicDeviceDriver
     {
-      //! Water velocity message.
-      IMC::GroundVelocity m_gvel;
+      //! Driver
+      std::unique_ptr<Driver> m_driver;
+      //! IO Handle
+      std::unique_ptr<IO::Handle> m_io;
       //! DVL altitude estimate.
       IMC::Distance m_alt_dvl;
       //! Filtered Altitude.
       IMC::Distance m_alt_flt;
       //! Beam Filter.
       Navigation::BeamFilter* m_filter;
-      //! True if using serial link.
-      bool m_serial;
-      //! Ethernet or Serial port handle.
-      IO::Handle* m_handle;
-      //! Message fragment buffer.
-      std::string m_msg_frag;
-      //! Timestamp of received data.
-      double m_timestamp;
       //! List of entities.
       std::vector<unsigned> m_entities;
-      //! Bottom lock.
-      bool m_bottom_lock;
-      //! Command ticket.
-      std::string m_cmd_ticket;
       //! Medium handler.
       DUNE::Monitors::MediumHandler m_hand;
       //! Task arguments.
       Arguments m_args;
-      //! Device configuration.
-      struct Configuration
-      {
-        double speed_of_sound;
-        bool acoustic_enabled;
-        bool dark_mode_enabled;
-        double mounting_rotation_offset;
-        std::string range_mode;
-      } m_config;
 
       //! Constructor.
       //! @param[in] name task name.
       //! @param[in] ctx context.
       Task(const std::string& name, Tasks::Context& ctx) :
-        Hardware::BasicDeviceDriver(name, ctx),
-        m_filter(NULL),
-        m_serial(false),
-        m_handle(NULL),
-        m_timestamp(-1),
-        m_bottom_lock(false)
+        Hardware::BasicDeviceDriver(name, ctx), m_filter(NULL)
       {
         param("IO Port - Device", m_args.io_dev)
             .defaultValue("")
             .description("IO device URI in the form \"tcp://ADDRESS:PORT\" "
                          "or \"uart://DEVICE:BAUD\".");
 
-        param("Power Channel - Names", m_args.pwr_channels)
-            .defaultValue("")
-            .description("Device's power channels");
+        param("Power Channel - Names", m_args.pwr_channels).defaultValue("").description("Device's power channels");
 
         param("Acoustics Activation", m_args.type_activation)
             .values("Water, Always")
@@ -221,22 +191,29 @@ namespace Sensors
       bool
       onConnect() override
       {
-        m_handle = openSocketTCP(m_args.io_dev);
-
-        if (m_handle == nullptr)
+        bool is_serial = false;
+        if (!m_driver)
         {
-          m_handle = openUART(m_args.io_dev);
-          m_serial = true;
+          m_driver = std::make_unique<Driver>(this, m_alt_dvl, m_alt_flt);
         }
+        m_io.reset(openSocketTCP(m_args.io_dev));
 
-        return m_handle != nullptr;
+        if (!m_io)
+        {
+          m_io.reset(openUART(m_args.io_dev));
+          is_serial = true;
+        }
+        m_driver->setIOHandle(m_io.get(), is_serial);
+
+        return m_io != nullptr;
       }
 
       //! Disconnect from device.
       void
       onDisconnect() override
       {
-        Memory::clear(m_handle);
+        m_io.reset();
+        m_driver.reset();
         Memory::clear(m_filter);
       }
 
@@ -250,12 +227,13 @@ namespace Sensors
                                   m_args.orientation, BeamFilter::CLOCKWISE);
 
         m_filter->setSourceEntities(m_entities);
+        m_driver->setBeamFilter(m_filter);
 
         // Initialize device configuration
-        getConfig();
-        setRotation(m_args.orientation[2]);
+        m_driver->getConfig();
+        m_driver->setRotation(m_args.orientation[2]);
         if (m_args.type_activation == "Always")
-          setAcoustics(true);
+          m_driver->setAcoustics(true);
       }
 
       void
@@ -270,209 +248,26 @@ namespace Sensors
           return;
 
         if (!m_hand.isKnown())
-          setAcoustics(false);
+          m_driver->setAcoustics(false);
         else if (m_hand.outWater())
-          setAcoustics(false);
+          m_driver->setAcoustics(false);
         else
-          setAcoustics(true);
-      }
-
-      //! Parser for incoming TPC data.
-      //! @param[in] data JSON data string
-      void
-      parseTCP(std::string& data)
-      {
-        try
-        {
-          // Interpret received data
-          auto msg = json::parse(data);
-
-          if (data.find("time") != std::string::npos)
-            parseTCPReport(msg);
-          else if (data.find("ts") != std::string::npos)
-            parseTCPDeadReckoning(msg);
-          else if (data.find("response_to") != std::string::npos)
-            parseTCPConfig(msg);
-          else
-            war("Unknown TCP message.");
-        }
-        catch (Exception &e)
-        {
-          err("TCP driver parsing problem '%s'.", e.what());
-          spew("\ntcp problematic parsed data = >>%s<<", Streams::sanitize(data).c_str());
-        }
-      }
-
-      //! Parse velocity-and-transducer report.
-      //! @param[in] msg parsed JSON data.
-      void
-      parseTCPReport(const nlohmann::json& msg)
-      {
-        if (!m_filter)
-          return;
-
-        // Status
-        uint8_t status = msg["status"];
-        checkStatus(status);
-
-        // Water Velocity
-        uint8_t validity = msg["velocity_valid"];
-        if (validity)
-        {
-          m_gvel.x = (fp64_t)msg["vx"];
-          m_gvel.y = (fp64_t)msg["vy"];
-          m_gvel.z = (fp64_t)msg["vz"];
-          m_gvel.validity =
-              (IMC::WaterVelocity::VAL_VEL_X | IMC::WaterVelocity::VAL_VEL_Y | IMC::WaterVelocity::VAL_VEL_Z);
-          m_bottom_lock = true;
-        }
-        else
-        {
-          m_bottom_lock = false;
-          m_gvel.clear();
-        }
-        m_gvel.setTimeStamp(m_timestamp);
-        dispatch(m_gvel, DF_KEEP_TIME);
-
-        // Bottom Ranges
-        uint8_t beam_valid;
-        float distance;
-        for (unsigned i = 0; i < c_beams; ++i)
-        {
-          beam_valid = msg["transducers"][i]["beam_valid"];
-          distance = (float)msg["transducers"][i]["distance"];
-
-          m_filter->setValidity(i, (beam_valid && distance > 0) ? IMC::Distance::DV_VALID : IMC::Distance::DV_INVALID);
-          m_filter->update(i, distance);
-        }
-        m_filter->dispatch(m_timestamp);
-
-        // Altitude from DVL
-        m_alt_dvl.value = (fp32_t)msg["altitude"];
-        m_alt_dvl.validity = validity;
-        m_alt_dvl.setTimeStamp(m_timestamp);
-        dispatch(m_alt_dvl, DF_KEEP_TIME);
-
-        // Filtered Altitude
-        m_alt_flt.value = m_filter->get();
-        m_alt_flt.validity = m_alt_flt.value > 0.0 ? IMC::Distance::DV_VALID : IMC::Distance::DV_INVALID;
-        m_alt_flt.setTimeStamp(m_timestamp);
-        dispatch(m_alt_flt, DF_KEEP_TIME);
-      }
-
-      //! Parse Dead Reckoning report.
-      void
-      parseTCPDeadReckoning(nlohmann::json& msg)
-      {
-        // Status
-        uint8_t status = msg["status"];
-        checkStatus(status);
-      }
-
-      //! Parse configuration response.
-      //! @param[in] msg parsed JSON data.
-      void
-      parseTCPConfig(const nlohmann::json& msg)
-      {
-        if ((std::string)msg["response_to"] == "get_config")
-        {
-          m_config.speed_of_sound = (double)msg["result"]["speed_of_sound"];
-          m_config.acoustic_enabled = (bool)msg["result"]["acoustic_enabled"];
-          m_config.dark_mode_enabled = (bool)msg["result"]["dark_mode_enabled"];
-          m_config.mounting_rotation_offset = (double)msg["result"]["mounting_rotation_offset"];
-          m_config.range_mode = (std::string)msg["result"]["range_mode"];
-        }
-
-        if (m_cmd_ticket.empty())
-          return;
-
-        if ((bool)msg["success"])
-        {
-          if ((std::string)msg["response_to"] == m_cmd_ticket)
-            m_cmd_ticket.clear();
-        }
-        else
-        {
-          std::string s(msg["error_message"]);
-          err("Configuration response error: %s", s.c_str());
-        }
-      }
-
-      //! Parser to interpret incoming data through serial connection.
-      //! @param[in] data JSON data string.
-      void
-      parseSerial(std::string& data)
-      {
-        war("Serial driver not supported yet.");
-        spew("\nSerial data = >>%s<<", Streams::sanitize(data).c_str());
-      }
-
-      //! Read data from sensor.
-      //! @return true if data is read, false otherwise.
-      bool
-      readData()
-      {
-        if (!Poll::poll(*m_handle, 0.01))
-          return false;
-
-        char bfr[2000] = { 0 };
-        IMC::DevDataText data;
-
-        size_t rv = m_handle->read(bfr, sizeof(bfr));
-        if (rv <= 0)
-          return false;
-        bfr[rv] = '\0';
-
-        m_timestamp = Clock::getSinceEpoch();
-
-        // Ensure proper string of data
-        std::vector<std::string> proper;
-        String::split(bfr, "\r\n", proper);
-        checkFragment(proper);
-        for (auto p : proper)
-        {
-          if (!m_serial)
-            parseTCP(p);
-          else
-            parseSerial(p);
-
-          // Log Raw Data
-          data.value = p.c_str();
-          dispatch(data);
-        }
-
-        return true;
-      }
-
-      //! Check for message fragments.
-      //! @param[in] list message list.
-      void
-      checkFragment(std::vector<std::string>& list)
-      {
-        if (!m_msg_frag.empty() && list.front().front() != '{')
-        {
-          list.front() = m_msg_frag + list.front();
-          m_msg_frag.clear();
-        }
-        if (list.back().find("}", list.back().find("json_v")) == std::string::npos)
-        {
-          m_msg_frag = list.back();
-          list.pop_back();
-        }
+          m_driver->setAcoustics(true);
       }
 
       //! Check for device faults.
       //! @param[in] status device status (0 -> OK; 1 -> fault).
       void
-      checkStatus(const bool status)
+      checkStatus()
       {
+        uint8_t status = m_driver->getLastStatus();
         if (status)
         {
           if (getEntityState() > IMC::EntityState::ESTA_NORMAL)
             return;
 
           // Disable acoustics until recovery
-          setAcoustics(false);
+          m_driver->setAcoustics(false);
           std::stringstream ss;
           ss << Status::getString(CODE_INTERNAL_ERROR) << " - DVL temperature too high"
              << " | Acoustics: Off";
@@ -486,28 +281,9 @@ namespace Sensors
             return;
 
           // Recovered from fault
-          setAcoustics(true);
+          m_driver->setAcoustics(true);
           displayConfig();
         }
-      }
-
-      //! Get device configuration.
-      void
-      getConfig(void)
-      {
-        sendCommand("get_config");
-      }
-
-      //! Enable or disable device acoustics.
-      //! @param[in] enable true to enable, false to disable.
-      void
-      setAcoustics(const bool enable)
-      {
-        if (m_config.acoustic_enabled == enable)
-          return;
-
-        sendCommand("set_config", "acoustic_enabled", enable ? "true" : "false");
-        getConfig();
       }
 
       //! Set device sound speed.
@@ -515,101 +291,7 @@ namespace Sensors
       void
       onSoundSpeed(const double sound_speed) override
       {
-        if (m_config.speed_of_sound == sound_speed || m_handle == nullptr)
-          return;
-        std::stringstream ss;
-        ss << sound_speed;
-        sendCommand("set_config", "speed_of_sound", ss.str());
-        getConfig();
-      }
-
-      //! Set range mode.
-      //! Modes table:
-      //!   Mode    Min alt (m)   Max alt (m)
-      //!   0       0.05          0.6
-      //!   1       0.3           3.0
-      //!   2       1.5           14
-      //!   3       7.7           36
-      //!   4       15            max
-      //! @param[in] spec Range specifier option:
-      //! auto  - The DVL will search for bottom lock in it's full operational area (Default)
-      //! =a    - The DVL is locked to range mode a where a is a number from 0-4
-      //! a<=b  - The DVL will search for bottom lock within range mode a and b
-      void
-      setRange(const std::string spec)
-      {
-        if (m_config.range_mode == spec)
-          return;
-
-        sendCommand("set_config", "range_mode", spec);
-        getConfig();
-      }
-
-      //! Set device rotation angle.
-      //! @param[in] angle mounting rotation angle in degrees.
-      void
-      setRotation(const double angle)
-      {
-        if (m_config.mounting_rotation_offset == angle)
-          return;
-
-        std::stringstream ss;
-        ss << angle;
-        sendCommand("set_config", "mounting_rotation_offset", ss.str());
-        getConfig();
-      }
-
-      //! Send command to device.
-      //! @param[in] cmd command type.
-      //! @param[in] param_name parameter name, only used for "set_config" command.
-      //!                       If empty only cmd is used.
-      //! @param[in] param_value parameter value, only used for "set_config" command.
-      //!                       If empty only cmd is used.
-      void
-      sendCommand(const std::string cmd, const std::string param_name = "", const std::string param_value = "")
-      {
-        if (!m_handle)
-          return;
-
-        std::string bfr;
-        if (!m_serial)
-        {
-          if (param_name.empty() || param_value.empty())
-            bfr = String::str("{\"command\":\"%s\"}", cmd.c_str());
-          else
-            bfr = String::str("{\"command\":\"%s\",\"parameters\":{\"%s\":%s}}", cmd.c_str(), param_name.c_str(),
-                              param_value.c_str());
-        }
-        else
-        {
-          war("Serial commands not implemented");
-          return;
-        }
-
-        spew("Sending: %s", bfr.c_str());
-        m_handle->writeString(bfr.c_str());
-
-        // Wait for response
-        if (!waitResponse(cmd))
-          war("%s failure (timeout)", cmd.c_str());
-      }
-
-      //! Wait command response
-      //! @return true if successful response, false otherwise
-      bool
-      waitResponse(const std::string& cmd)
-      {
-        m_cmd_ticket = cmd;
-        double ttl = Clock::get() + c_ttl;
-        while (Clock::get() < ttl)
-        {
-          readData();
-          if (m_cmd_ticket.empty())
-            return true;
-        }
-
-        m_cmd_ticket.clear();
-        return false;
+        m_driver->setSoundSpeed(sound_speed);
       }
 
       //! Get data from device.
@@ -617,9 +299,9 @@ namespace Sensors
       bool
       onReadData() override
       {
-        bool recv = readData();
+        bool recv = m_driver->readData();
+        checkStatus();
         displayConfig();
-
         return recv;
       }
 
@@ -627,10 +309,12 @@ namespace Sensors
       void
       displayConfig(void)
       {
+        Configuration config = m_driver->getLastConfig();
+        bool bottom_lock = m_driver->getBottomLock();
         std::string description;
         description = String::str("%s - Acoustics: %s | Sound Speed: %.1f m/s",
-                                  Status::getString(m_bottom_lock ? CODE_ACTIVE : CODE_NO_BOTTOM_LOCK),
-                                  m_config.acoustic_enabled ? "On" : "Off", m_config.speed_of_sound);
+                                  Status::getString(bottom_lock ? CODE_ACTIVE : CODE_NO_BOTTOM_LOCK),
+                                  config.acoustic_enabled ? "On" : "Off", config.speed_of_sound);
         setEntityState(getEntityState(), description);
       }
     };
